+++ srcs/int/ft_intlen.c
@@ -6,13 +6,13 @@
 /*   By: wbraeckm <wbraeckm@student.s19.be>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/06/14 09:15:17 by wbraeckm          #+#    #+#             */
-/*   Updated: 2018/06/14 09:21:15 by wbraeckm         ###   ########.fr       */
+/*   Updated: 2018/07/13 13:40:49 by wbraeckm         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_intlen(int value)
+size_t	ft_intlen(ssize_t value)
 {
 	return (ft_intlen_base(value, 10));
 }
