+++ srcs/int/ft_intlen_base.c
@@ -6,13 +6,13 @@
 /*   By: wbraeckm <wbraeckm@student.s19.be>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/06/14 09:20:07 by wbraeckm          #+#    #+#             */
-/*   Updated: 2018/06/18 00:25:52 by wbraeckm         ###   ########.fr       */
+/*   Updated: 2018/07/13 13:45:24 by wbraeckm         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_intlen_base(int value, size_t basesize)
+size_t	ft_intlen_base(ssize_t value, size_t basesize)
 {
 	size_t i;
 
@@ -21,7 +21,7 @@ size_t	ft_intlen_base(int value, size_t basesize)
 	i = 0;
 	while (value != 0)
 	{
-		value /= (int)basesize;
+		value /= (ssize_t)basesize;
 		i++;
 	}
 	return (i);
